functions/currentTemp/currentTemp/requirements.txt
google-cloud-pubsub==0.39.1
flask
pymongo
redis
